avoid `registerCompletionItemProvider` run multi times
