+++ modules/front-end/src/app/core/components/webhook-drawer/webhook-drawer.component.ts
@@ -11,8 +11,7 @@ import { NzMessageService } from "ng-zorro-antd/message";
 import { ProjectService } from "@services/project.service";
 import { IEnvironment, IProject } from "@shared/types";
 import { of } from "rxjs";
-// eslint-disable-next-line  @typescript-eslint/no-explicit-any
-declare const monaco: any;
+import { MonacoService } from "@services/monaco-service";
 
 @Component({
   selector: 'webhook-drawer',
@@ -39,7 +38,8 @@ export class WebhookDrawerComponent implements OnInit {
     private fb: FormBuilder,
     private projectService: ProjectService,
     private webhookService: WebhookService,
-    private message: NzMessageService
+    private message: NzMessageService,
+    private monacoService: MonacoService
   ) {
     this.initForm();
   }
@@ -99,46 +99,7 @@ export class WebhookDrawerComponent implements OnInit {
   };
 
   onEditorInit(editor: editor.IStandaloneCodeEditor): void {
-    // register editor variables
-    function createDependencyProposals(range) {
-      // returning a static list of proposals, not even looking at the prefix (filtering is done by the Monaco editor),
-      // here you could do a server side lookup
-      return [
-        {
-          label: '"@@flag.name"',
-          kind: monaco.languages.CompletionItemKind.Variable,
-          documentation: "The name of the feature flag",
-          insertText: '@@flag.name',
-          range: range
-        },
-        {
-          label: '"@@flag.description"',
-          kind: monaco.languages.CompletionItemKind.Variable,
-          documentation: "The description of the feature flag",
-          insertText: '@@flag.description',
-          range: range
-        }
-      ];
-    }
-
-    monaco.languages.registerCompletionItemProvider("json", {
-      provideCompletionItems: function (model, position) {
-        // Get the text before the cursor
-        const word = model.getWordUntilPosition(position);
-
-        const idx = word.word.lastIndexOf('@');
-        const range = {
-          startLineNumber: position.lineNumber,
-          endLineNumber: position.lineNumber,
-          startColumn: word.startColumn + Math.max(0, idx),
-          endColumn: word.endColumn,
-        };
-
-        return { suggestions: createDependencyProposals(range) };
-      },
-
-      triggerCharacters: ['@']
-    });
+    this.monacoService.init();
 
     // format the document
     setTimeout(() => {
